Move VoiceInstructionConfig to seperate file
